+++ mitra_config/src/federation.rs
@@ -11,7 +11,6 @@ const fn default_deliverer_pool_size() -> usize { 10 }
 const fn default_deliverer_timeout() -> u64 { 10 }
 const fn default_deliverer_log_response_length() -> usize { 75 }
 const fn default_fep_1b12_full_enabled() -> bool { true }
-const fn default_fep_c0e0_emoji_react_enabled() -> bool { false }
 const fn default_fep_ef61_gateway_enabled() -> bool { false }
 const fn default_fep_171b_public_enabled() -> bool { false }
 
@@ -55,9 +54,6 @@ pub struct FederationConfig {
     )]
     pub fep_1b12_full_enabled: bool,
 
-    #[serde(default = "default_fep_c0e0_emoji_react_enabled")]
-    pub fep_c0e0_emoji_react_enabled: bool,
-
     #[serde(default = "default_fep_ef61_gateway_enabled")]
     pub fep_ef61_gateway_enabled: bool,
 
@@ -82,7 +78,6 @@ impl Default for FederationConfig {
             onion_proxy_url: None,
             i2p_proxy_url: None,
             fep_1b12_full_enabled: default_fep_1b12_full_enabled(),
-            fep_c0e0_emoji_react_enabled: default_fep_c0e0_emoji_react_enabled(),
             fep_ef61_gateway_enabled: default_fep_ef61_gateway_enabled(),
             fep_171b_public_enabled: default_fep_171b_public_enabled(),
         }