+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Changed the default value of `limits.posts.attachment_local_limit` to `4`.
 
+### Removed
+
+- Removed `federation.fep_c0e0_emoji_react_enabled` configuration parameter.
+
 ## [4.20.0] - 2026-03-19
 
 ### Added